Add option to prevent the form to be submitted when hitting ENTER
Goal: add an option to avoid the form to be submitted when hitting
ENTER on the places autocompleter

New property preventSubmit added. Set to false by default to not change
the default behavior.

Add a new keyDown event handler to prevent the ENTER key to submit the
form if the option is enabled and we have a location.

Let me know if you need more information.